test(User) create abstraction to user tests.
